Flyttet lasting av saksopplysninger ut til Saksbehandling.
