+++ src/sider/saksbehandling.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
 import { Container, Row, Column } from 'nav-frontend-grid';
 import './saksbehandling.css';
 import Vilkarsvurdering from '../felles-komponenter/vilkarsvurdering/vilkarsvurdering';
@@ -9,10 +10,12 @@ import Medlemskap from '../felles-komponenter/medlemskap';
 import SideOppsummering from '../felles-komponenter/sideOppsummering';
 import SideDialog from '../felles-komponenter/sideDialog/sideDialog';
 import SideKommentarer from '../felles-komponenter/sideKommentarer';
+import { hentSaksopplysninger } from '../ducks/saksopplysninger';
 
 class Saksbehandling extends Component {
   componentDidMount() {
-
+    const { saksnr } = this.props.match.params;
+    this.props.hentSaksopplysninger(saksnr);
   }
 
   render() {
@@ -41,6 +44,8 @@ class Saksbehandling extends Component {
 
 const mapStateToProps = () => ({});
 
-const mapDispatchToProps = () => ({});
+const mapDispatchToProps = dispatch => ({
+  hentSaksopplysninger: saksnr => dispatch(hentSaksopplysninger(saksnr)),
+});
 
-export default connect(mapStateToProps, mapDispatchToProps)(Saksbehandling);
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Saksbehandling));
